chore(suite-native): reorganize connect files
