+++ suite-native/module-connect-device/src/components/connect/ConnectDeviceScreenHeader.tsx
