+++ suite-native/module-connect-device/src/components/connect/ConnectDeviceSreenView.tsx
