+++ suite-native/module-connect-device/src/components/connect/PinFormProgress.tsx
