+++ suite-native/module-connect-device/src/components/connect/PinOnDevice.tsx
@@ -18,10 +18,10 @@ import TrezorConnect, { UI_REQUEST } from '@trezor/connect';
 const SCREEN_HEIGHT = Dimensions.get('screen').height;
 
 export const deviceImageMap: Record<DeviceModelInternal, string> = {
-    [DeviceModelInternal.T1B1]: require('../assets/pin-t1b1.png'),
-    [DeviceModelInternal.T2T1]: require('../assets/pin-t2t1.png'),
-    [DeviceModelInternal.T2B1]: require('../assets/pin-t2b1.png'),
-    [DeviceModelInternal.T3T1]: require('../assets/pin-t3t1.png'),
+    [DeviceModelInternal.T1B1]: require('../../assets/pin-t1b1.png'),
+    [DeviceModelInternal.T2T1]: require('../../assets/pin-t2t1.png'),
+    [DeviceModelInternal.T2B1]: require('../../assets/pin-t2b1.png'),
+    [DeviceModelInternal.T3T1]: require('../../assets/pin-t3t1.png'),
 };
 
 const wrapperStyle = prepareNativeStyle(utils => ({
